requirements.txt
# Numpy>=1.26.4
# Pandas>=2.3.2
# Jupyter>=1.1.1
# JupyterLab>=4.4.7
# mysql-connector-python>=9.4.0
# matplotlib>=3.10.6
# seaborn>=0.13.2
# openpyxl>=3.1.5
# peewee>=3.18.2
# psycopg2-binary>=2.9.11
# PyMySQL>=1.1.2
# python-decouple>=3.8
# SQLAlchemy>=2.0.44

numpy==1.26.4
pandas==2.3.2
jupyter==1.1.1
jupyterlab==4.4.7
notebook==7.4.5
mysql-connector-python==9.4.0
matplotlib==3.10.6
seaborn==0.13.2
openpyxl==3.1.5
scikit-learn==1.7.1
scipy==1.16.1
peewee==3.18.2
psycopg2-binary==2.9.11
PyMySQL==1.1.2
python-decouple==3.8
SQLAlchemy==2.0.44
